requirements.txt
dj-database-url==0.4.2
Django==2.0
django-cors-headers==2.1.0
djangorestframework==3.7.3
djangorestframework-jwt==1.11.0
geopy==1.11.0
gunicorn==19.7.1
psycopg2==2.7.3.2
PyJWT==1.5.3
pytz==2017.3
